chore: add more subnets (#212)
This PR adds more subnets, which allows us to update cluster endpoint access and give the dev team access to the EKS cluster.

Due to using consecutive numbers in existing private and public subnets I've bumped the next private subnet to `105`. IDK, I wanted 5, but it was not available so I added 100.