fix: update argo-cd to v2.5 for arm support (#222)
the argocd pods are crashing as the older version is not built for arm.

We seem to be stuck with 20 bitswap peer pods, as that was the last
config where argocd worked. I have not been able to scale up that number
manually using kubectl.

Also the cluster autoscaler appears to be scaling the number of sever
instances to the minimum number requires to fit the pods. we asked for
14 server instances but it scaled it down to 7. each bitswap peer asks
for 4GB ram, and each instance has ~12GB available for bitswap pods.

So we can fit up to 3 bs pods per instance so it gives us 7 intances.

from the numbers I think we need to scale the number of bs pods back up
to ~30. The cluster scaler should do it's thing and provision ~10
intances again.

BUT. We it seems we can't scale that number without fixing argocd. Note
that there are 2 argo tools installed here "argo-cd" which I think
controls the scaling of pods & "argo-rollouts" which controls canary
deployments. (But argo-rollouts is confusingly labelled
argocd-rollouts). That important note is that our "argo-rollouts" pods
*are working* but our argo-cd pods are crash looping due to running an
old version.

So we can rollout new versions of the container image but we seem to be
unable to scale the number of the pods.



see: https://github.com/argoproj/argo-cd/issues/9121

License: MIT

---------

Signed-off-by: Oli Evans <oli@protocol.ai>